feat: extract duplicated monaco webpack config into build-config
