+++ client/build-config/src/index.ts
@@ -1,2 +1,3 @@
 export * from './webpack/getCssLoaders'
 export * from './paths'
+export * from './webpack/monaco-editor'
