+++ client/build-config/src/paths.ts
@@ -1,3 +1,4 @@
 import path from 'path'
 
 export const ROOT_PATH = path.resolve(__dirname, '../../../')
+export const NODE_MODULES_PATH = path.resolve(ROOT_PATH, 'node_modules')
