+++ client/web/dev/esbuild/monacoPlugin.ts
@@ -4,8 +4,9 @@ import * as esbuild from 'esbuild'
 import { EditorFeature, featuresArr } from 'monaco-editor-webpack-plugin/out/features'
 import { EditorLanguage, languagesArr } from 'monaco-editor-webpack-plugin/out/languages'
 
+import { MONACO_LANGUAGES_AND_FEATURES } from '@sourcegraph/build-config'
+
 import { ROOT_PATH } from '../utils'
-import { MONACO_LANGUAGES_AND_FEATURES } from '../webpack/monacoWebpack'
 
 const monacoModulePath = (modulePath: string): string =>
     require.resolve(path.join('monaco-editor/esm', modulePath), {
