--- client/web/dev/webpack/monacoWebpack.ts
@@ -1,33 +1,0 @@
-import type MonacoWebpackPlugin from 'monaco-editor-webpack-plugin'
-
-/**
- * Configuration for https://github.com/microsoft/monaco-editor-webpack-plugin.
- */
-export const MONACO_LANGUAGES_AND_FEATURES: Required<
-    Pick<
-        NonNullable<ConstructorParameters<typeof MonacoWebpackPlugin>[0]>,
-        'languages' | 'features' | 'customLanguages'
-    >
-> = {
-    languages: ['json', 'yaml'],
-    customLanguages: [
-        {
-            label: 'yaml',
-            entry: 'monaco-yaml/lib/esm/monaco.contribution',
-            worker: { id: 'vs/language/yaml/yamlWorker', entry: 'monaco-yaml/lib/esm/yaml.worker' },
-        },
-    ],
-    features: [
-        'bracketMatching',
-        'clipboard',
-        'coreCommands',
-        'cursorUndo',
-        'find',
-        'format',
-        'hover',
-        'inPlaceReplace',
-        'iPadShowKeyboard',
-        'links',
-        'suggest',
-    ],
-}
